Move the package from io.github.slimeyar to io.github.slimeymc
