--- common/src/main/java/io/github/slimeyar/slimeys_utility/integration/SlimeysUtilityKubeJSPlugin.java
@@ -1,16 +1,0 @@
-package io.github.slimeyar.slimeys_utility.integration;
-
-import dev.latvian.mods.kubejs.script.BindingsEvent;
-import io.github.slimeyar.slimeys_utility.physicify.ShipPhysicifyKt;
-import io.github.slimeyar.slimeys_utility.util.VectorConversionKt;
-
-
-public class SlimeysUtilityKubeJSPlugin {
-    public static void addBindings(BindingsEvent event) {
-        // Vector Conversion
-        event.add("VectorConversionsExt", VectorConversionKt.class);
-
-        // ShipPhysicify
-        event.add("ShipPhysicify", ShipPhysicifyKt.class);
-    }
-}
