--- common/src/main/kotlin/io/github/slimeyar/slimeys_utility/SlimeysUtilityMod.kt
@@ -1,18 +1,0 @@
-package io.github.slimeyar.slimeys_utility
-
-import org.slf4j.LoggerFactory
-
-object SlimeysUtilityMod {
-
-    const val RENDER_TYPE = "slimeys_utility_render"
-
-    const val MOD_ID = "slimeys_utility"
-
-    val LOGGER = LoggerFactory.getLogger(MOD_ID)
-
-    @JvmStatic
-    fun init() {
-        SlimeysUtilityItems.register()
-        SlimeysUtilityNetworking.registerServer()
-    }
-}
