+++ common/src/main/resources/slimeys_utility-common.mixins.json
@@ -1,10 +1,10 @@
 {
   "required": true,
-  "package": "io.github.slimeyar.slimeys_utility.mixin",
+  "package": "io.github.slimeymc.slimeys_utility.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "features.GameRendererMixin",
-    "features.ShipAssemblyMixin"
+    "io.github.slimeymc.slimeys_utility.mixin.features.GameRendererMixin",
+    "io.github.slimeymc.slimeys_utility.mixin.features.ShipAssemblyMixin"
   ],
   "injectors": {
     "defaultRequire": 1
