--- fabric/src/main/java/io/github/slimeyar/slimeys_utility/fabric/ValkyrienPreLaunch.java
@@ -1,12 +1,0 @@
-package io.github.slimeyar.slimeys_utility.fabric;
-
-import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
-
-/**
- * For now, just using this class as an abusive early entrypoint to run the updater
- */
-public class ValkyrienPreLaunch implements PreLaunchEntrypoint {
-
-    @Override
-    public void onPreLaunch() {}
-}
