+++ forge/src/main/resources/slimeys_utility.mixins.json
@@ -2,7 +2,7 @@
   "required": true,
   "package": "io.github.priestoffern.vs_ship_assembler.forge.mixin",
   "compatibilityLevel": "JAVA_17",
-  "plugin": "io.github.slimeyar.slimeys_utility.forge.mixin.ValkyrienForgeMixinConfigPlugin",
+  "plugin": "io.github.slimeymc.slimeys_utility.forge.mixin.ValkyrienForgeMixinConfigPlugin",
   "mixins": [],
   "injectors": {
     "defaultRequire": 1
